fix: refactoring based on feedback
