+++ src/experimentation-ess/common.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Adobe. All rights reserved.
+ * Copyright 2024 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -13,7 +13,6 @@
 /* c8 ignore start */
 import RUMAPIClient from '@adobe/spacecat-shared-rum-api-client';
 import { JSDOM } from 'jsdom';
-import { AuditBuilder } from '../common/audit-builder.js';
 import { getRUMDomainkey } from '../support/utils.js';
 
 const EXPERIMENT_PLUGIN_OPTIONS = {
@@ -22,7 +21,7 @@ const EXPERIMENT_PLUGIN_OPTIONS = {
   experimentsMetaTag: 'experiment',
   experimentsQueryParameter: 'experiment',
 };
-const DAYS = 30;
+
 const METRIC_CHECKPOINTS = ['click', 'convert', 'formsubmit'];
 
 let log = console;
@@ -428,6 +427,7 @@ async function processExperimentRUMData(experimentInsights) {
 }
 
 export async function processAudit(auditURL, context, site, days) {
+  log = context.log;
   const rumAPIClient = RUMAPIClient.createFrom(context);
   const domainkey = await getRUMDomainkey(site.getBaseURL(), context);
   const options = {
@@ -437,35 +437,8 @@ export async function processAudit(auditURL, context, site, days) {
     granularity: 'hourly',
   };
   const experimentData = await rumAPIClient.query('experiment', options);
-  log.info(`Experiment Insighsts for ${DAYS} day(s): ${JSON.stringify(experimentData, null, 2)}`);
+  log.info(`Experiment Insighsts for ${days} day(s): ${JSON.stringify(experimentData, null, 2)}`);
   return processExperimentRUMData(experimentData);
 }
 
-export async function essExperimentationAuditRunner(auditUrl, context, site) {
-  log = context.log;
-  log.info(`Received ESS Experimentation Monthly audit request for ${auditUrl}`);
-  const startTime = process.hrtime();
-
-  const auditData = await processAudit(
-    auditUrl,
-    context,
-    site,
-    DAYS,
-  );
-
-  const endTime = process.hrtime(startTime);
-  const elapsedSeconds = endTime[0] + endTime[1] / 1e9;
-  const formattedElapsed = elapsedSeconds.toFixed(2);
-
-  log.info(`ESS Experimentation Monthly Audit completed in ${formattedElapsed} seconds for ${auditUrl}`);
-
-  return {
-    auditResult: auditData,
-    fullAuditRef: auditUrl,
-  };
-}
-
-export default new AuditBuilder()
-  .withRunner(essExperimentationAuditRunner)
-  .build();
 /* c8 ignore stop */
