+++ src/index.js
@@ -26,8 +26,8 @@ import sitemap from './sitemap/handler.js';
 import backlinks from './backlinks/handler.js';
 import experimentation from './experimentation/handler.js';
 import conversion from './conversion/handler.js';
-import essExperimentationMonthly from './ess-experimentation-monthly/handler.js';
-import essExperimentationDaily from './ess-experimentation-daily/handler.js';
+import essExperimentationMonthly from './experimentation-ess/monthly.js';
+import essExperimentationDaily from './experimentation-ess/daily.js';
 
 const HANDLERS = {
   apex,
